# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SIAA — Sistema Integrado de Asistencia Académica
  version: 0.1.0
  description: |
    API REST de SIAA. Todos los endpoints requieren HTTPS. Las respuestas de
    error siguen la estructura del catálogo §9.2 del SRS.
  contact:
    name: Equipo SIAA
    email: dev@siaa.edu.co
  license:
    name: Propietario
servers:
  - url: https://api.siaa.edu.co/api/v1
    description: Producción
  - url: http://localhost:8080/api/v1
    description: Desarrollo local

# ─── Seguridad global ──────────────────────────────────────────────────────
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Esquemas de error comunes ─────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required: [codigo, mensaje, correlationId]
      properties:
        codigo:
          type: string
          example: AUTH_CREDENCIALES_INVALIDAS
          description: Código de error del catálogo §9.2
        mensaje:
          type: string
          example: Credenciales incorrectas
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
        correlationId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
    FieldError:
      type: object
      required: [campo, error]
      properties:
        campo:
          type: string
          example: correo
        error:
          type: string
          example: required

    # ─── Auth ──────────────────────────────────────────────────────────
...
    RoleDTO:
      type: object
      required: [nombre, permisos]
      properties:
        nombre:
          type: string
          example: DOCENTE
        permisos:
          type: array
          items:
            type: string
          example: [marcaje:crear, marcaje:leer, horario:leer]

    # ─── Jerarquía Física y Espacios (US-GEO-01) ─────────────────
    SedeDTO:
      type: object
      required: [id, codigo, nombre, activo, creadoEn, actualizadoEn]
      properties:
        id:
          type: string
          example: 66f0a1b2c3d4e5f60718293a
        codigo:
          type: string
          example: SEDE-CENTRAL
        nombre:
          type: string
          example: Campus Principal
        direccion:
          type: string
          example: Calle 10 # 5-20
        activo:
          type: boolean
          example: true
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
    CrearSedeRequest:
      type: object
      required: [codigo, nombre]
      properties:
        codigo:
          type: string
          example: SEDE-CENTRAL
        nombre:
          type: string
          example: Campus Principal
        direccion:
          type: string
          example: Calle 10 # 5-20
    BloqueDTO:
      type: object
      required: [id, sedeId, codigo, nombre, activo, creadoEn, actualizadoEn]
      properties:
        id:
          type: string
          example: 66f0a1b2c3d4e5f60718293b
        sedeId:
          type: string
          example: 66f0a1b2c3d4e5f60718293a
        codigo:
          type: string
          example: B-ING
        nombre:
          type: string
          example: Bloque de Ingeniería
        pisos:
          type: array
          items:
            type: integer
          example: [1, 2, 3, 4]
        activo:
          type: boolean
          example: true
        creadoEn:
          type: string
          format: date-time
        actualizadoEn:
          type: string
          format: date-time
    CrearBloqueRequest:
      type: object
      required: [sedeId, codigo, nombre]
      properties:
        sedeId:
          type: string
          example: 66f0a1b2c3d4e5f60718293a
        codigo:
          type: string
          example: B-ING
        nombre:
          type: string
          example: Bloque de Ingeniería
        pisos:
          type: array
          items:
            type: integer
          example: [1, 2, 3, 4]
...
paths:
...
  /roles:
    get:
      operationId: listarRoles
      summary: Listar roles del sistema
      description: |
        Retorna la lista de roles predefinidos con sus permisos según la matriz del SRS §3.2 (AC-01 US-ROL-01).
        Requiere token de acceso con permiso rol:leer.
      tags: [roles]
      responses:
        '200':
          description: Lista de roles obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere rol:leer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: crearRolPersonalizado
      summary: Crear rol personalizado (US-ROL-02)
      tags: [roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, permisos]
              properties:
                nombre:
                  type: string
                permisos:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Rol creado exitosamente

  # ─── Sedes (US-GEO-01) ───────────────────────────────────────────────────────
  /sedes:
    get:
      operationId: listarSedes
      summary: Listar todas las sedes institucionales activas
      tags: [sedes]
      responses:
        '200':
          description: Lista de sedes obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SedeDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere aula:leer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: crearSede
      summary: Registrar una nueva sede
      tags: [sedes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSedeRequest'
      responses:
        '201':
          description: Sede creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SedeDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere sede:administrar)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto de unicidad de código
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sedes/{id}:
    get:
      operationId: obtenerSede
      summary: Obtener detalle de una sede por ID
      tags: [sedes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sede obtenida correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SedeDTO'
        '404':
          description: Sede no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Bloques (US-GEO-01) ────────────────────────────────────────────────────
  /bloques:
    get:
      operationId: listarBloques
      summary: Listar bloques (opcionalmente filtrados por sedeId)
      tags: [bloques]
      parameters:
        - name: sedeId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de bloques obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BloqueDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere aula:leer)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: crearBloque
      summary: Registrar un nuevo bloque en una sede
      tags: [bloques]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearBloqueRequest'
      responses:
        '201':
          description: Bloque creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloqueDTO'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permiso denegado (requiere bloque:administrar)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto de unicidad de código
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /bloques/{id}:
    get:
      operationId: obtenerBloque
      summary: Obtener detalle de un bloque por ID
      tags: [bloques]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bloque obtenido correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BloqueDTO'
        '404':
          description: Bloque no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Espacios (US-GEO-01) ───────────────────────────────────────────────────
...
  /roles/{id}:
    get:
      operationId: obtenerRolPorId
      summary: Obtener detalle de un rol (US-ROL-02)
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle del rol
    put:
      operationId: actualizarRol
      summary: Actualizar rol personalizado (US-ROL-02)
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rol actualizado
    delete:
      operationId: eliminarRol
      summary: Eliminar rol personalizado (US-ROL-02)
      tags: [roles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Rol eliminado
...